fixed name null, started ingredient unit seeder

+++ database/migrations/2023_05_03_210124_create_ingredient_recipe_unit_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('recipe_id')->index();
             $table->foreignId('ingredient_id')->index();
             $table->foreignId('unit_id')->index()->nullable();
-            $table->integer('value');
+            $table->float('value');
 
             $table->timestamps();
         });
